Catch up to main branch (#4)
* Repair selection of palette section for ColorFromOverlayForm

Repair selection of palette section for ColorFromOverlayForm by re-connecting the selected index changed event handler to the palette section combo box of the ColorFromOverlayForm.

Re-colored design will appear in the main window when OK for the form is selected. This does not include any use of the preview panel.

* Implemented preview in ColorFromOverlayForm

Provide a preview image in the ColorFromOverlyForm when the dialog is initialized, when the palette section is changed, and when the ring size filter is changed.

* Changes to the ColorFromOverlayForm and its handling

Implemented the ability to use the L*, a*, and b* sliders and checkboxes to adjust the colors computed from the overlay prior to matching them to the palette colors.

Added a progress bar so that there is some indication that something is happening as the preview image is being computed.

Draw the preview image only once during form initialization, and delay the draw until the (blank) form is initially shown.  This provides a faster response to the menu choice having been selected, and also allows the user to see the progress bar move as the initial preview image is computed.

* Improve efficiency of ColorFromOverlayForm preview image generation.

Improved the efficiency of preview image generation in the ColorFromOverlayForm by computing the average overlay color for each ring only once, when the form is first shown.  This operation was by far the longest-running component of preview image generation.

Refactored re-generation of the preview image (and also the final coloring of the design) to  apply the ring size filter, selected palette section, and color transformation to the already-computed average color information.

* Version 3.2 and documentation update for color-from-overlay feature.

Updated version to 3.2.
Updated help and user manual for changes to the color-from-overlay feature.

---------

Co-authored-by: Christopher M. Albrecht <76419094+CMAlbrecht@users.noreply.github.com>